lambda_functions/document_verifier/requirements.txt
PyPDF2>=3.0.0
python-docx>=1.1.0
Pillow>=10.0.0
pytesseract>=0.3.10
